Fix bug where setting reminder.notified to True and saving it overwrote deletions or edits
Moved setting reminder.notified to True to chat_completion_from_reminder() in chat/models.py because the chat completion can delete or edit the reminder, and so if we do it in tasks.py then we can potentially be working on a reminder object that no longer exists in the database, and therefore be overwriting any adjustments or deletions. Plus, it's more logical to change reminder.notified to True right after adding the notification message.