Dockerize application
Create the docker-compose.yml that runs four containers: web, celery, celerybeat, and redis. Change all secrets to be stored in environmental variables because Azure key vault doesn't work well with Docker containers unless the Docker containers are running in Azure (which will be the case in production). Comment out production code, including the Azure key vault code and the code to send messages to Twilio.