Migrate to Discord chat bot, add TODOs
Fully migrated to Discord chat bot. The chat bot runs from a script using a websocket. The websocket then sends requests to the Django application for each message. Redis is used as a message broker to send messages between Celery and the chat bot when reminders are triggered. This is because the websocket cannot take HTTP requests.
Cleanup for residual Twilio code is still required.